+++ 01_krepios/lang/de/init.lua
@@ -29,6 +29,7 @@ local myPackages = {
 
     "01_krepios/lang/de/side_missions/de_buyer.lua",
     "01_krepios/lang/de/side_missions/de_capture.lua",
+    "01_krepios/lang/de/side_missions/de_destroy_graveyard.lua",
     "01_krepios/lang/de/side_missions/de_disable_ship.lua",
     "01_krepios/lang/de/side_missions/de_gather_crystals.lua",
     "01_krepios/lang/de/side_missions/de_pirate_base.lua",