+++ 01_krepios/lang/en/init.lua
@@ -29,6 +29,7 @@ local myPackages = {
 
     "01_krepios/lang/en/side_missions/en_buyer.lua",
     "01_krepios/lang/en/side_missions/en_capture.lua",
+    "01_krepios/lang/en/side_missions/en_destroy_graveyard.lua",
     "01_krepios/lang/en/side_missions/en_disable_ship.lua",
     "01_krepios/lang/en/side_missions/en_gather_crystals.lua",
     "01_krepios/lang/en/side_missions/en_pirate_base.lua",